chore: updated

+++ src/Mud.Core.Abstractions/Models/DailyInfo.cs
@@ -3,14 +3,14 @@
 public class DailyInfo
 {
     public DateTime Date { get; set; }
-    public decimal High { get; set; }
-    public decimal Open { get; set; }
-    public decimal Close { get; set; }
-    public decimal Low { get; set; }
+    public double High { get; set; }
+    public double Open { get; set; }
+    public double Close { get; set; }
+    public double Low { get; set; }
 
-    public decimal Volume { get; set; }
+    public double Volume { get; set; }
 
-    public decimal PreClose { get; set; }
+    public double PreClose { get; set; }
 
-    public decimal Change { get; set; }
+    public double Change { get; set; }
 }
